enhance(backend): 凍結の後処理をQueueで処理するように
