+++ packages/backend/src/core/QueueService.ts
@@ -365,6 +365,16 @@ export class QueueService {
 	}
 
 	@bindThis
+	public createUserSuspendJob(user: ThinUser) {
+		return this.dbQueue.add('userSuspend', {
+			user: { id: user.id },
+		}, {
+			removeOnComplete: true,
+			removeOnFail: true,
+		});
+	}
+
+	@bindThis
 	public createReportAbuseJob(report: MiAbuseUserReport) {
 		return this.dbQueue.add('reportAbuse', report);
 	}
